Sudoku

Description

Documentation

data Cell

Constructors

Fixed Int

Possible [Int]

type Row = [Cell]

type Grid = [Row]

getCellData :: Char -> Maybe Cell

The getCellData function checks for a Fixed number, if not, a Possible 1-9 list is assigned

makeGrid :: String -> Maybe Grid

The makeGrid function creates a grid by running getCellData over every char in a string

reducePossibilities :: [Cell] -> Maybe [Cell]

Checks for Fixed values in a row, takes the Fixed number out of every possibility in that row

s :: [Char]

grid :: Grid

grid' :: [[Cell]]

grid'' :: [[Cell]]

main :: IO ()

The main prints the grid with reduced possibilities in Rows and Columns